contacts v0.1.0 Contacts.Queries

Perform the queries to database

Functions

create_contact(attrs)

create_contact(map()) :: {:ok, Ecto.Schema.t()} | {:error, Ecto.Changeset.t()}

Create a contact given a map of attrs

Examples

create_contact(%{name: "John", email: "john@example.com", surname: "Doe", phone_number: "54912345678"})

get_contact(email)

get_contact(binary()) :: Ecto.Schema.t() | nil | no_return()

Return the contact specified by email

Example

get_contact("Jhon@example.com")

list_contact(params)

list_contact(map()) :: [Ecto.Schema.t()] | no_return()

Return a list of contacts. Accept params to filter and sort

Params

field name: filter the search by field value

_ord: allow to order by asc or desc, default: asc

_sort: allow to specify a field to sort, default: email

Example

list_contact(%{"_ord" => "desc", "_sort" => "name", "surname" => "Doe"})

mark_as_delete(email)

mark_as_delete(binary()) ::
  {:ok, Ecto.Schema.t()} | nil | {:error, Ecto.Changeset.t()}

Mark the contact with specific email to be removed by the garbage collector

Examples

mark_as_delete("john@example.com")

update_contact(email, attrs)

update_contact(binary(), map()) ::
  {:ok, Ecto.Schema.t()} | nil | {:error, Ecto.Changeset.t()}

Update the contact with the email specified with the mapattrs

Examples

update_contact("john@example.com", %{name: "Cameron", phone_number: "+54912345678"})